test(/fake): Fake 객체들을 한 패키지에 모음
