+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -7,8 +7,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import org.junit.jupiter.api.Test;
-import roomescape.FakeReservationDaoImpl;
-import roomescape.FakeReservationTimeDaoImpl;
+import roomescape.fake.FakeReservationDaoImpl;
+import roomescape.fake.FakeReservationTimeDaoImpl;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.repository.ReservationDao;
